Moved tests to test directory (and renamed the files).
