--- examples/test/Makefile
@@ -1,2 +1,0 @@
-all:
-	g++ Main.cpp TestCase.cpp -std=c++11 -Wall -lpredict -lm -o test
