--- examples/test/README
@@ -1,1 +1,0 @@
-Usage: ./test <testfiles>
