--- examples/test/testcase
@@ -1,60 +1,0 @@
-[tle]
-1 25544U 98067A   15129.86961041  .00015753  00000-0  23097-3 0  9998
-2 25544  51.6464 275.3867 0006524 289.1638 208.5861 15.55704207942078
-
-[qth]
-lat=63.9
-lon=10.9
-alt=0.0
-
-[data]
-13025.788888888888323,-47.77631081,112.67812075,402.593,106.755,-67.487
-13025.788900462963284,-47.75092055,112.76251738,402.579,106.667,-67.495
-13025.788912037036425,-47.72546191,112.84682813,402.565,106.579,-67.503
-13025.788923611111386,-47.69993504,112.93105291,402.551,106.492,-67.511
-13025.788935185184528,-47.67433903,113.01519504,402.537,106.404,-67.519
-13025.788946759259488,-47.64867605,113.09924769,402.523,106.316,-67.527
-13025.788958333332630,-47.62294522,113.18321418,402.509,106.228,-67.535
-13025.788969907407591,-47.59714665,113.26709443,402.495,106.140,-67.543
-13025.788981481480732,-47.57128046,113.35088840,402.481,106.053,-67.550
-13025.788993055555693,-47.54534679,113.43459601,402.467,105.965,-67.558
-13025.789004629628835,-47.51934577,113.51821721,402.453,105.876,-67.566
-13025.789016203703795,-47.49327750,113.60175194,402.438,105.788,-67.573
-13025.789027777776937,-47.46714108,113.68520348,402.424,105.700,-67.581
-13025.789039351851898,-47.44093872,113.76856508,402.410,105.612,-67.589
-13025.789050925925039,-47.41466951,113.85184004,402.396,105.524,-67.596
-13025.789062500000000,-47.38833357,113.93502829,402.381,105.435,-67.604
-13025.789074074074961,-47.36193102,114.01812979,402.367,105.347,-67.611
-13025.789085648148102,-47.33546199,114.10114449,402.353,105.259,-67.618
-13025.789097222223063,-47.30892661,114.18407233,402.338,105.170,-67.626
-13025.789108796296205,-47.28232393,114.26691660,402.324,105.082,-67.633
-13025.789120370371165,-47.25565621,114.34967058,402.310,104.993,-67.640
-13025.789131944444307,-47.22892252,114.43233756,402.295,104.905,-67.648
-13025.789143518519268,-47.20212298,114.51491750,402.281,104.816,-67.655
-13025.789155092592409,-47.17525772,114.59741035,402.266,104.727,-67.662
-13025.789166666667370,-47.14832685,114.67981606,402.252,104.639,-67.669
-13025.789178240740512,-47.12133051,114.76213461,402.237,104.550,-67.676
-13025.789189814815472,-47.09426883,114.84436593,402.222,104.461,-67.683
-13025.789201388888614,-47.06714083,114.92651331,402.208,104.372,-67.690
-13025.789212962963575,-47.03994883,115.00857009,402.193,104.283,-67.697
-13025.789224537036716,-47.01269185,115.09053954,402.179,104.194,-67.704
-13025.789236111111677,-46.98537003,115.17242162,402.164,104.105,-67.711
-13025.789247685184819,-46.95798350,115.25421630,402.149,104.016,-67.718
-13025.789259259259779,-46.93053237,115.33592355,402.134,103.927,-67.725
-13025.789270833332921,-46.90301677,115.41754334,402.120,103.838,-67.732
-13025.789282407407882,-46.87543572,115.49907891,402.105,103.749,-67.738
-13025.789293981481023,-46.84779157,115.58052367,402.090,103.659,-67.745
-13025.789305555555984,-46.82008332,115.66188087,402.075,103.570,-67.752
-13025.789317129629126,-46.79231111,115.74315050,402.060,103.481,-67.758
-13025.789328703704086,-46.76447507,115.82433251,402.046,103.391,-67.765
-13025.789340277777228,-46.73657531,115.90542690,402.031,103.302,-67.771
-13025.789351851852189,-46.70861196,115.98643362,402.016,103.212,-67.778
-13025.789363425925330,-46.68058402,116.06735592,402.001,103.123,-67.784
-13025.789375000000291,-46.65249387,116.14818726,401.986,103.033,-67.791
-13025.789386574073433,-46.62434051,116.22893088,401.971,102.944,-67.797
-13025.789398148148393,-46.59612406,116.30958675,401.956,102.854,-67.804
-13025.789409722221535,-46.56784465,116.39015487,401.941,102.764,-67.810
-13025.789421296296496,-46.53950241,116.47063521,401.926,102.675,-67.816
-13025.789432870369637,-46.51109746,116.55102776,401.911,102.585,-67.822
-13025.789444444444598,-46.48262992,116.63133250,401.895,102.495,-67.829
-13025.789456018517740,-46.45409878,116.71155264,401.880,102.405,-67.835
